API V1.0
Router : add path functionnalities
API : basic api able to do everything for the project. some safeguards have not yet been implemented and should be in a future version

+++ api/index.php
@@ -8,7 +8,7 @@
 *   /user/{id} get the json output of the database for the user at the id {id}
 *   /user/{id}/projects get the projects the user has access to.
 *   /user/me  shortcut to /user/{my id} for authentified users.
-*   /user/add adds a user : it neeed POST parameters
+*   /user/new adds a user : it neeed POST parameters
 *      -name
 *      -email
 *      -password
@@ -62,33 +62,66 @@
 **/
 
 
-require_once "routage.php";
+require_once "router.php";
 require_once "../functions.php";
 
-if(!isset($_COOKIE["PHPSESSID"]))
-{
+if (session_status() == PHP_SESSION_NONE) {
     session_start();
 }
 
 header('Content-type: application/json');
 
-function get($str){
+function get($str, $optionnal = false){
     if (isset($_GET[$str])){
         return $_GET[$str];
     }
 
-    die("missing GET parameter - $str");
+    if (! $optionnal){
+        die("missing GET parameter - $str");
+    }
+
+    return null;
 }
 
-function post($str){
+function post($str, $optionnal = false){
     if (isset($_POST[$str])){
         return $_POST[$str];
     }
+    if (! $optionnal){
+        die("missing POST parameter - $str");
+    }
 
-    die("missing POST parameter - $str");
+    return null;
 }
 
+function error_401(){
+    header("Content-Type: text/plain");
+    header("HTTP/1.1 401 Unauthorized");
+    echo "You must authentificate to access this service";
+    exit;
+}
 
+function error_403($msg="Access Forbidden"){
+    header("Content-Type: text/plain");
+    header("HTTP/1.1 403 Forbidden");
+    echo $msg;
+    exit;
+}
+
+/**
+* return the user id if he is connected
+* kill the process otherwise
+**/
+function force_auth(){
+    if (isset($_SESSION["user_id"])){
+        return $_SESSION["user_id"];
+    }else{
+        die("error - you need to log in to continue");
+    }
+}
+
+$route = new Route();
+function route(...$args){global $route;$route->route(...$args);};
 
 /*-----------------------------------------------------------------------------------------------------------------------------------------*/
 
@@ -99,18 +132,23 @@ function post($str){
 * /disconnext disconnect a user - clear all cookies and delete the session
 *
 **/
-route("/api/login", function(){
+$route->post(array("/api/login",
+                   "/api/user/login"), function(){
     $mail = post("email");
-    $password = hash_passwd(post("password"));
+    $password = post("password");
 
     if (validate_user($mail, $password)){
         $_SESSION["user_id"] = get_user_by_email($mail)["id"];
-
-
+        echo "user logged in";
+    }else{    
+        echo "login failed";
     }
 });
 
-route("/api/disconnect", function(){
+$route->route(array("/api/logout",
+                    "/api/disconnect",
+                    "/api/user/logout",
+                    "/api/user/disconnect"), function(){
     if (session_status() == PHP_SESSION_ACTIVE) { session_destroy(); }
 });
 
@@ -127,16 +165,15 @@ route("/api/disconnect", function(){
 *      -email
 *      -password
 **/
-
-//todo : check this function
-route("/api/user/add", function(){
+$route->post(array("/api/user/new",
+                   "/api/user/add"), function(){
     $name = post("name");
     $email = post("email");
     $password = post("password");
 
     if (user_test_mail($email)){
         //todo : add 409 error
-        die();
+        die("error - mail already exists");
     }
 
     $id_user = add_user($name, $email, $password);
@@ -147,19 +184,81 @@ route("/api/user/add", function(){
 
 //todo : check this function
 route("/api/user/me", function(){    
-    $id = $_SESSION["user_id"];
+    $id = force_auth();
 
     $output = get_user($id);
     echo json_encode($output);
 });
 
-route("/api/user/{id}", function($id){
+$route->get("/api/user/{id}", function($id){
     $id = (int) $id;
     $output = get_user($id);
     echo json_encode($output);
 });
 
-route("/api/user/{id}/projects", function($id){
+/** edit the user info - POST
+* POST: user_id
+* optional parameter : password, name, email
+*
+* @return the user infos
+**/
+//todo : add a security
+$route->post("/api/user/{id}/edit", function($id){
+    $id = (int) $id;
+    $name = post("name", true);
+    $email = post("email", true);
+    $password = post("password", true);
+
+    $args = array(":id"=>$id);
+    $set = array();
+
+    if ($name){
+        $args[":name"] = $name;
+        $set[]="name = :name";
+    }
+
+    if ($email){
+        $args[":email"] = $email;
+        $set[]="email = :email";
+    }
+
+    if ($password){
+        $args[":password"] = hash_passwd($password);
+        $set[]="password = :password";
+    }
+
+    if (count($set) >= 1){
+        $req = "UPDATE users SET ".join(",",$set)." WHERE id=:id";
+        execute($req, $args);
+    }
+
+
+    $output = get_user($id);
+    echo json_encode($output);
+});
+
+$route->delete(array("/api/user/me/delete",
+                     "/api/user/{id}/delete"),
+               function($id = 0){
+                   $id = $id !== 0 ? (int) $id : force_auth();
+
+                   //todo : add more security
+                   if (isset($_SESSION["user_id"]) && $id == $_SESSION["user_id"]){
+                       delete_user($id);
+                       if ($id == $_SESSION["user_id"]){
+                           session_destroy();
+                       }
+                   }else{
+                       echo $id." - ".$_SESSION["user_id"]."\n";
+                       echo "you cannot destroy an account if it is not yours / you are not connected";
+                   }
+
+                   $output = array();
+                   echo json_encode($output);
+               });
+
+//todo check here
+$route->route("/api/user/{id}/projects", function($id){
     $id = (int) $id;
     $output = get_projects_for_user($id);
     echo json_encode($output);
@@ -188,27 +287,74 @@ route("/api/user/{id}/projects", function($id){
 *
 *   /project/{id}/tickets return tickets of the project
 *   /project/{id}/ticket/{id_simple_ticket} return the ticket
-*   /project/add add a project; POST parameters
+*   /project/new add a project; POST parameters
 *      -name
 *      -ticket_prefix
 *      
 *      returns th project as it is in the db
 **/
 
-route("/api/project/{id}", function($id){
+$route->post("/api/project/new", function(){
+    $name = post("name");
+    $ticket_prefix = post("ticket_prefix");
+    $id_user = force_auth();
+
+    $id_project = add_project($name,$id_user, $ticket_prefix);
+
+    $output = array("id_project"=>$id_project);
+    echo json_encode($output);
+});
+
+
+$route->get("/api/project/{id}", function($id){
     $id = (int) $id;
     $project = get_project($id);
     echo json_encode($project);
 });
 
-route("/api/project/{id}/delete", function($id){
-    $id = (int) $id;
-    //todo : add verification that the logged user CAN do it
-    delete_project($id);
+/*only the creator can change these parameters for now*/
+$route->post("/api/project/{id}/edit", function($id){
+    $name = post("name", true);
+    $ticket_prefix = post("ticket_prefix", true);
+    $id_user = force_auth();
+
+    $args = array(":id"=>$id,
+                  ":creator_id"=>$id_user);
+    $set = array();
+
+    if ($name){
+        $args[":name"] = $name;
+        $set[]="name = :name";
+    }
+
+    if ($ticket_prefix){
+        $args[":ticket_prefix"] = $ticket_prefix;
+        $set[]="ticket_prefix = :ticket_prefix";
+    }
+
+    if (count($set) >= 1){
+        $req = "UPDATE projects SET ".join(",",$set)." WHERE id=:id AND creator_id=:creator_id;";
+        execute($req, $args);
+    }
+
+    $output = get_project($id);
+    echo json_encode($output);
 });
 
-//todo : check this function
-route("/api/project/{id}/adduser", function($id_project){
+/*only the creator can delete for now*/
+$route->delete("/api/project/{id}/delete", function($id){
+    $id = (int) $id;
+    $id_user = force_auth();
+
+    if (is_admin($id_user,$id)){
+        delete_project($id);
+    }else{
+        echo "error - you do not have the right access level to do that";
+    }
+});
+
+//todo : add user verification and add a return
+$route->post("/api/project/{id}/adduser", function($id_project){
     $id_project = (int) $id_project;
     //todo : add verification that the logged user CAN do it
     $id_user = (int) post("id_user");
@@ -217,8 +363,15 @@ route("/api/project/{id}/adduser", function($id_project){
     add_link_user_project($id_user, $id_project, $access_level);    
 });
 
-//todo : check this function
-route("/api/project/{id}/removeuser", function($id_project){
+/*return the users on the project*/
+//todo : verify if the user has the right to see that
+$route->get("/api/project/{id}/users", function($id){
+    $id = (int) $id;
+    echo json_encode(get_users_for_project($id));
+});
+
+//todo : verify the user CAN do it
+$route->post("/api/project/{id}/removeuser", function($id_project){
     $id_project = (int) $id_project;
     //todo : add verification that the logged user CAN do it
     $id_user = (int) post("id_user");
@@ -226,19 +379,143 @@ route("/api/project/{id}/removeuser", function($id_project){
     delete_link_user_project($id_user, $id_project);
 });
 
-route("/api/project/{id}/addticket", function($id_project){
+//todo : verify theuser can do it
+$route->post("/api/project/{id}/addticket", function($id_project){
     $id_project = (int) $id_project;
     $title = post("title");
     $priority = post("priority");
     $description = post("description");
     $due_date = post("due_date");
 
-    $creator_id = $_SESSION["user_id"];
+    $creator_id = force_auth();
 
+    $id = add_ticket($title, $id_project, $creator_id, $creator_id, $priority, $description, $due_date);
+
+    $output = array("id_ticket" => $id);
+    echo json_encode($output);    
+});
+
+//todo : add false case and vverify if user has the right to check this
+$route->get("/api/project/{id}/tickets", function($id_project){
+    echo json_encode(get_tickets_for_project($id_project));
+});
+
+//todo : add false case
+$route->get("/api/project/{id_project}/ticket/{id_simple_ticket}", function($id_project, $id_simple_ticket){
+    echo json_encode(get_ticket_simple($id_project, $id_simple_ticket));
 });
 
 
+/**
+* TICKETS
+*   /ticket/{id} return the ticket information IF the user has access to the project
+                 equivalent to /project/{id}/ticket/{id_simple_ticket}. all the following can be used on both path
+*   /ticket/{id}/comments get the comments of the ticket
+*      -comment
+*   /ticket/{id}/addcomment POST
+*      -comment
+*      user needs to be authenticated
+**/
+
+$route->get("/api/ticket/{id}", function($id){
+    echo json_encode(get_ticket($id));
+});
+
+/*todo : check the user can do that*/
+$route->post("/api/ticket/{id}/edit", function($id_ticket){
+    $id_ticket = (int) $id_ticket;
+    $params = array(
+        "name"=>post("name", true),
+        "priority"=>post("priority", true),
+        "description"=>post("description", true),
+        "due_date"=>post("due_date", true)
+    );
+
+    $args = array(":ticket_id"=>$id_ticket);
+    $set = array();
+
+    foreach($params as $t=>$v){
+        if ($v){
+            $args[":$t"] = $v;
+            $set[] = "$t = :$t";
+        }
+    }
+
+    if (count($set) >= 1){
+        $req = "UPDATE tickets SET ".join(",",$set)." WHERE id=:ticket_id;";
+        execute($req, $args);
+    }
+
+    echo json_encode(get_ticket($id_ticket));
+});
+
+/*todo : check the user can do that*/
+$route->delete("/api/ticket/{id}/delete", function($id_ticket){
+    delete_ticket($id_ticket);
+});
+
+/*todo : check the user can do that AND that the ticket exists*/
+$route->post("/api/ticket/{id}/addcomment", function($ticket_id){
+    $ticket_id = (int) $ticket_id;
+    $comment = post("comment");
+    $creator_id = force_auth();
+
+    $id = add_comment($ticket_id, $creator_id, $comment);
+    $output = array("id_comment"=>$id);
+    echo json_encode($output);
+});
+
+$route->get("/api/ticket/{id}/comments", function($id){
+    echo json_encode(get_comments_for_ticket($id));
+});
 
 
+/**
+* COMMENTS
+* GET /api/comment/{id}
+* POST /api/comment/{id}/edit
+*    -comment
+* DELETE /api/comment/{id}/delete
+*
+**/
+
+//todo : check if the user has rights
+$route->get("/api/comment/{id_comment}", function($id_comment){
+    echo json_encode(get_comment($id_comment));
+});
+
+$route->post("/api/comment/{id}/edit", function($comment_id){
+    $comment = post("comment");
+    $creator_id = force_auth();
+
+    if (get_comment($comment_id)["creator_id"] != $creator_id){
+        die("You cannot modify a comment that isn't yours");
+    }
+
+    $id = edit_comment($comment_id, $comment);
+
+    echo json_encode(get_comment($comment_id));
+});
+
+$route->delete("/api/comment/{id}/delete", function($comment_id){
+    $comment_id = (int) $comment_id;
+    $id_user = force_auth();
+    $id_project = execute("SELECT project_id FROM tickets WHERE id IN (SELECT ticket_id FROM comments WHERE id = ?)", array($comment_id))->fetch()["project_id"];
+    
+    if (is_admin($id_user, $id_project) || get_comment($comment_id)["creator_id"] == $creator_id){
+        delete_comment($comment_id);
+    }
+});
+
+/**
+* Error Handling
+*
+*
+**/
+
+$route->error_404(function(){
+    echo "Oops... It seems i was unable to do anything with the request you gave me."."\n<br/>";
+    echo "Maybe check the type and the url ?";
+});
 
 ?>
